# docs/api/openapi.public.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpsNinja API",
    "version": "1.0.0",
    "description": "OpsNinja is a multi-tenant SaaS support and incident-management platform. This document describes the versioned, publicly supported tenant-facing REST surface. Internal endpoints (health, auth, admin, Jira integration) are excluded from this document.",
    "contact": {
      "name": "OpsNinja API Support",
      "url": "https://docs.opsninja.io",
      "email": "api-support@opsninja.io"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://opsninja.io/terms"
    }
  },
  "servers": [
    { "url": "https://api.opsninja.io", "description": "Production" },
    { "url": "https://api.staging.opsninja.io", "description": "Staging" },
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
...
  "security": [{ "StaffBearer": [] }],
  "paths": {
...
    "/api/v1/portal/attachments/{id}/download": {
      "get": {
        "operationId": "downloadPortalAttachment",
        "summary": "Get a pre-signed download URL for a portal attachment",
        "tags": ["portal-attachments"],
        "security": [{ "PortalBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Pre-signed download URL valid for 5 minutes.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AttachmentDownload" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/portal/attachments/presign": {
      "post": {
        "operationId": "presignPortalAttachment",
        "summary": "Obtain a pre-signed S3 POST URL for portal attachment upload",
        "description": "Returns a pre-signed S3 POST URL and form fields. The client POSTs the file directly to S3 (multipart/form-data), then calls the confirm endpoint. The OpsNinja API never receives the file body.",
        "tags": ["portal-attachments"],
        "security": [{ "PortalBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PresignAttachmentRequest" },
              "example": { "filename": "screenshot.png", "mimeType": "image/png", "sizeBytes": 204800 }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pre-signed upload URL and attachment record.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PresignAttachmentResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/api/v1/portal/attachments/{id}/confirm": {
      "post": {
        "operationId": "confirmPortalAttachment",
        "summary": "Mark a portal attachment upload as complete",
        "tags": ["portal-attachments"],
        "security": [{ "PortalBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Attachment confirmed and ready to be linked to a ticket.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PortalAttachmentMeta" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
    "/api/v1/tickets/{id}/attachments/presign": {
      "post": {
        "operationId": "presignAttachment",
        "summary": "Obtain a pre-signed S3 POST URL for an agent attachment upload",
        "description": "Returns a pre-signed URL and form fields for direct S3 upload. The API server never receives the file body (AC: file upload contract).",
        "tags": ["agent-attachments"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PresignAttachmentRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Pre-signed upload URL.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PresignAttachmentResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
    "/api/v1/attachments/{id}/download": {
      "get": {
        "operationId": "downloadAttachment",
        "summary": "Get a pre-signed download URL for an agent attachment",
        "tags": ["agent-attachments"],
        "security": [{ "StaffBearer": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Pre-signed download URL valid for 5 minutes.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AttachmentDownload" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": true,
        "description": "Field-level or context detail item.",
        "example": { "field": "subject", "issue": "String must contain at least 1 character(s)" }
      },
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string", "description": "Machine-readable error code.", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "description": "Human-readable error description.", "example": "Validation failed" },
          "details": { "type": "array", "items": { "$ref": "#/components/schemas/ErrorDetail" } },
          "traceId": { "type": "string", "example": "01HV2MXPQ3Y5HRGZTBJE8DKNN4" }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorBody" }
        },
        "example": {
          "error": {
            "code": "NOT_FOUND",
            "message": "Ticket not found.",
            "details": [],
            "traceId": "01HV2MXPQ3Y5HRGZTBJE8DKNN4"
          }
        }
      },
...
      "PortalAttachmentMeta": {
        "type": "object",
        "required": ["id", "displayName", "mimeType", "sizeBytes"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "displayName": { "type": "string" },
          "mimeType": { "type": "string" },
          "sizeBytes": { "type": "integer" }
        }
      },
...
      "AttachmentDownload": {
        "type": "object",
        "required": ["url", "expiresAt"],
        "properties": {
          "url": { "type": "string", "format": "uri" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
      "PresignAttachmentRequest": {
        "type": "object",
        "required": ["filename", "mimeType", "sizeBytes"],
        "additionalProperties": false,
        "properties": {
          "filename": { "type": "string", "maxLength": 255 },
          "mimeType": { "type": "string" },
          "sizeBytes": { "type": "integer", "minimum": 1, "maximum": 26214400 }
        }
      },
      "PresignAttachmentResponse": {
        "type": "object",
        "required": ["attachmentId", "uploadUrl", "fields", "expiresAt"],
        "properties": {
          "attachmentId": { "type": "string", "format": "uuid" },
          "uploadUrl": { "type": "string", "format": "uri" },
          "fields": { "type": "object", "additionalProperties": { "type": "string" } },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "Bad Request — validation failed or malformed input.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid bearer token.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "Forbidden": {
        "description": "Forbidden — insufficient permissions.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
      "NotFound": {
        "description": "Not Found — resource does not exist or is out of scope.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      },
...
      "TooManyRequests": {
        "description": "Too Many Requests — rate limit exceeded.",
        "headers": {
          "Retry-After": {
            "description": "Number of seconds to wait before retrying.",
            "schema": { "type": "integer", "example": 60 }
          }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } }
      }
    },
    "securitySchemes": {
      "StaffBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC JWT issued by the tenant identity provider (Okta / Entra ID). Staff and agent principals."
      },
      "PortalBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Portal-scoped JWT bound to a single customer organisation."
      },
      "MachineToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Long-lived tenant-issued API key for machine-to-machine integrations."
      }
    }
  }
}